Refactor import statement in certbot-new.py to use letsencrypt module for improved clarity
